tech: refactor wine and proton methods to another file
